[JSSC-67] 🔨 Refactor : UploadImageTitle 컴포넌트 분리
